# openapi.yaml
openapi: 3.0.1
info:
  title: James' Bin
  version: '1'
  description: An e-commerce website for spy craft goods.
paths:
  /login:
    post:
      summary: Login.
      description: Login using username and password.
      operationId: post_login
      requestBody:
        description: login object with username and password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login'
      responses:
        '200':
          description: Success. Redirect.
        default:
          description: Default error sample response
      tags:
        - Authentication
  /logout:
    post:
      summary: Logout.
      description: Logout. No input required.
      operationId: post_logout
      responses:
        '200':
          description: Success. Redirect.
        default:
          description: Default error sample response
      tags:
        - Authentication
  /register:
    post:
      summary: Creates a new user.
      description: Creats a new user using a request object. Hashes and salts the password in the object.
      operationId: register
      requestBody:
        description: User object.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/register'
      responses:
        '200':
          description: Success. Redirect.
        default:
          description: Default error sample response
      tags:
        - Authentication
  /user/password:
    put:
      summary: Change password.
      description: Change password with hash and salt.
      operationId: put_user_password
      requestBody:
        description: Object with password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/password'
      responses:
        '200':
          content:
            text/plain; charset=utf-8:
              examples:
                users_id:
                  value: Password updated.
          description: Success. Redirect.
        default:
          description: Default error sample response
      tags:
        - Authentication
...
components:
  schemas:
    login:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      xml:
        name: login
    register:
      type: object
      properties:
        fullname:
          type: string
        username:
          type: string
        password:
          type: string
        contact:
          type: object
          properties:
            phone:
              type: string
            address:
              type: string
            city:
              type: string
            state:
              type: string
            zip:
              type: string
            email:
              type: string
      xml:
        name: user
    password:
      type: object
      properties:
        password:
          type: string
      xml:
        name: password
...